Remove installation of two scripts from CMakeLists.txt. Move particle_restart.py to utils/openmc folder to make it part of Python API.
