+++ tests/test_particle_restart_eigval/results.py
@@ -5,7 +5,7 @@ import sys
 sys.path.insert(0, '../../src/utils')
 
 # import particle restart
-import particle_restart as pr
+import openmc.particle_restart as pr
 
 # read in particle restart file
 if len(sys.argv) > 1:
