+++ tests/test_particle_restart_fixed/results.py
@@ -4,8 +4,8 @@ import sys
 
 sys.path.insert(0, '../../src/utils')
 
-# import particle restart 
-import particle_restart as pr
+# import particle restart
+import openmc.particle_restart as pr
 
 # read in particle restart file
 if len(sys.argv) > 1:
@@ -15,8 +15,8 @@ else:
 
 # set up output string
 outstr = ''
- 
-# write out properties 
+
+# write out properties
 outstr += 'current batch:\n'
 outstr += "{0:12.6E}\n".format(p.current_batch)
 outstr += 'current gen:\n'
